+++ src/usr/tst2/main.h
@@ -3,6 +3,6 @@
 
 void uprintf(char* fmt, ...);
 void uprintfstrcol_scr(unsigned char font, char* fmt);
-char * itoa(unsigned int n, unsigned int base);
+char * itoa(unsigned long n, unsigned int base);
 int main(void);
 #endif
